Makefile: Removed "-Werror" as a default flag.
"-Werror" is not a good idea for release builds as library function
may become deprecated over time and the project is no longer actively
maintained. Even for debug builds you often do not want to deal with
compiler warnings as quite often you just want to try if some
functionality is working now.